edit 3 (main_handler)
