+++ index.html
@@ -27,10 +27,10 @@
   <!--<script src="https://cdnjs.cloudflare.com/ajax/libs/crypto-js/4.1.1/crypto-js.js" integrity="sha512-NQVmLzNy4Lr5QTrmXvq/WzTMUnRHmv7nyIT/M6LyGPBS+TIeRxZ+YQaqWxjpRpvRMQSuYPQURZz/+pLi81xXeA==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>-->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/crypto-js/4.1.1/crypto-js.min.js" integrity="sha512-E8QSvWZ0eCLGk4km3hxSsNmGWbLtSCSUcewDQPQWZF6pEU8GlT8a5fF32wOl1i8ftdMhssTrF/OhyGWwonTcXA==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
   <script src="agora-rtm-sdk-1.5.1.js" ></script>
-  <script type="module" src="./test.js" ></script>
   <script type="module" src="./hash_handler.js" ></script>
   <script type="module" src="./sten.js" ></script>
   <script type="module" src="./rsa_handler.js"></script>
+  <script type="module" src="./main_handler.js" ></script>
   <script type="module" src="./main.js" ></script>
   
 </html>
