+++ main_handler.js
@@ -1,3 +1,4 @@
+import {encode as encode_arr, insertflag, decode as decode_arr,next_signed,next,distance} from './sten.js' 
 let width;
 let height;
 let inputCtx
@@ -51,7 +52,40 @@ let sendFrame = async (arr,dataChannel) => {
 
 }
 
+let sendMessage = async (text,inputCtx,width,height,dataChannel,remote_hash_str)=>{
+  
+  let index = 0
+  let encode_res; 
+  let temp_text = text; 
+  let first_time = true
+  let id=0
+  let part=0;
+  do{
+    await sleep(60);
+    inputCtx.drawImage(document.getElementById("user-1"), 0, 0, width, height);
+    let pixelData = inputCtx.getImageData(0, 0, width, height);
+    let arr = pixelData.data;
+    encode_res = encode_arr(arr,temp_text,id,part,remote_hash_str)
+    if(encode_res==-1){
+      continue;
+    }
+    index++
+    temp_text = encode_res.str
+    part++;
+    id = encode_res.id;
+    await sendFrame(arr,dataChannel);
+    
+  }
+  while(temp_text!="")
+
+  await dataChannel.send("end-message")
+  let convoTextarea = document.getElementById("callTextarea")
 
 
+  //writes the user part
+  convoTextarea.value +="me:\n"+text+'\n';
 
-export {makeid,sleep,str2frame,sendFrame}
+}
+
+
+export {makeid,sleep,str2frame,sendFrame, sendMessage}
